1355: Add middleware for healthcheck responses
