+++ accessibility_monitoring_platform/apps/common/tests/test_middleware.py
@@ -7,7 +7,7 @@ from django.core.exceptions import DisallowedHost
 from django.urls import reverse
 
 from ..models import UserCacheUniqueHash
-from ..middleware.validate_host_middleware import ValidateHostMiddleware
+from ..middleware.healthcheck_middleware import HealthcheckMiddleware
 
 
 @pytest.mark.django_db
@@ -20,18 +20,28 @@ def test_middleware_caches_user_unique_id(client, django_user_model):
     assert UserCacheUniqueHash.objects.all().count() == 1
 
 
-@patch(
-    "accessibility_monitoring_platform.apps.common.middleware.validate_host_middleware.logger.info"
-)
-def test_validate_host_middleware_valid_host(mock_logger):
-    """Tests host validation for valid host"""
-    host: str = settings.ALLOWED_HOSTS[0]
+def test_healthcheck_middleware_healthcheck_request():
+    """Tests healthcheck requests don't call get_response"""
     mock_request = Mock()
-    mock_request._get_raw_host.return_value = host
-    validate_host_middleware = ValidateHostMiddleware(Mock())
-    validate_host_middleware(mock_request)
+    mock_request.META = {"PATH_INFO": "/healthcheck/"}
+    mock_get_response = Mock()
+    healthcheck_middleware = HealthcheckMiddleware(Mock())
+    healthcheck_middleware.get_response = mock_get_response
+    healthcheck_middleware(mock_request)
 
-    mock_logger.assert_called_once_with("Valid host found: %s", host)
+    mock_get_response.assert_not_called()
+
+
+def test_healthcheck_middleware_non_healthcheck_request():
+    """Tests non-healthcheck requests call get_response"""
+    mock_request = Mock()
+    mock_request.META = {"PATH_INFO": "/other/"}
+    mock_get_response = Mock()
+    healthcheck_middleware = HealthcheckMiddleware(Mock())
+    healthcheck_middleware.get_response = mock_get_response
+    healthcheck_middleware(mock_request)
+
+    mock_get_response.assert_called_once()
 
 
 @pytest.mark.parametrize(
